Move BasicJobConfig.scala to correct package
